refactor: componentização do header @DhieRodrigues @ccarolb
